lib: Suppress comparison warnings in ovsdb libraries.
This patch fixes compiler warnings like the following:

./lib/ovsdb-types.h:171:5: error: comparison of unsigned expression
>= 0 is always true [-Werror=type-limits]

+++ lib/ovsdb-parser.c
@@ -80,7 +80,7 @@ ovsdb_parser_member(struct ovsdb_parser *parser, const char *name,
         return NULL;
     }
 
-    if ((value->type >= 0 && value->type < JSON_N_TYPES
+    if (((int) value->type >= 0 && value->type < JSON_N_TYPES
          && types & (1u << value->type))
         || (types & OP_ID && value->type == JSON_STRING
             && ovsdb_parser_is_id(value->u.string)))
